Clean up the ETM tests now that there's one implementation
