+++ tests/algorithms/extended_temporal_memory/etm_apical_tiebreak_test.py
@@ -20,18 +20,18 @@
 # ----------------------------------------------------------------------
 
 """
-Run the apical tiebreak tests on the C++ ExtendedTemporalMemory.
+Run the apical tiebreak tests on the ExtendedTemporalMemory.
 """
 
 import unittest
 
-from htmresearch_core.experimental import ExtendedTemporalMemory
+from nupic.bindings.experimental import ExtendedTemporalMemory
 from htmresearch.support.shared_tests.apical_tiebreak_test_base import (
   ApicalTiebreakTestBase)
 
 
-class ExtendedTMCPP_ApicalTiebreakTests(ApicalTiebreakTestBase,
-                                        unittest.TestCase):
+class ExtendedTM_ApicalTiebreakTests(ApicalTiebreakTestBase,
+                                     unittest.TestCase):
   """
   Run the apical tiebreak tests on the C++ ExtendedTemporalMemory.
   """
