+++ tests/algorithms/extended_temporal_memory/etm_sequence_memory_test.py
@@ -1,6 +1,6 @@
 # ----------------------------------------------------------------------
 # Numenta Platform for Intelligent Computing (NuPIC)
-# Copyright (C) 2017, Numenta, Inc.  Unless you have an agreement
+# Copyright (C) 2016-2017, Numenta, Inc.  Unless you have an agreement
 # with Numenta, Inc., for a separate license for this software code, the
 # following terms and conditions apply:
 #
@@ -20,30 +20,29 @@
 # ----------------------------------------------------------------------
 
 """
-Run the apical tiebreak sequence tests on the C++ ExtendedTemporalMemory.
+Run the sequence memory tests on the C++ ExtendedTemporalMemory.
 """
 
 import unittest
 
-from htmresearch_core.experimental import ExtendedTemporalMemory
-from htmresearch.support.shared_tests.apical_tiebreak_sequences_test_base import (
-  ApicalTiebreakSequencesTestBase)
+from nupic.bindings.experimental import ExtendedTemporalMemory
+
+from htmresearch.support.shared_tests.sequence_memory_test_base import (
+  SequenceMemoryTestBase)
 
 
-
-class ExtendedTMCPP_ApicalTiebreakSequencesTests(ApicalTiebreakSequencesTestBase,
-                                                 unittest.TestCase):
+class ExtendedTM_SequenceMemoryTests(SequenceMemoryTestBase,
+                                     unittest.TestCase):
   """
-  Run the apical tiebreak sequence tests on the C++ ExtendedTemporalMemory.
+  Run the sequence memory tests on the C++ ExtendedTemporalMemory.
   """
 
-  def constructTM(self, columnCount, apicalInputSize, cellsPerColumn,
-                  initialPermanence, connectedPermanence, minThreshold,
-                  sampleSize, permanenceIncrement, permanenceDecrement,
+  def constructTM(self, columnCount, cellsPerColumn, initialPermanence,
+                  connectedPermanence, minThreshold, sampleSize,
+                  permanenceIncrement, permanenceDecrement,
                   predictedSegmentDecrement, activationThreshold, seed):
 
     params = {
-      "apicalInputDimensions": (apicalInputSize,),
       "columnDimensions": (columnCount,),
       "cellsPerColumn": cellsPerColumn,
       "initialPermanence": initialPermanence,
@@ -56,25 +55,16 @@ class ExtendedTMCPP_ApicalTiebreakSequencesTests(ApicalTiebreakSequencesTestBase
       "activationThreshold": activationThreshold,
       "seed": seed,
       "learnOnOneCell": False,
-      "formInternalBasalConnections": True,
     }
 
     self.tm = ExtendedTemporalMemory(**params)
 
 
-  def compute(self, activeColumns, apicalInput, learn):
-
-    activeColumns = sorted(activeColumns)
-    apicalInput = sorted(apicalInput)
-
+  def compute(self, activeColumns, learn):
     # Use depolarizeCells + activateCells rather than tm.compute so that
     # getPredictiveCells returns predictions for the current timestep.
-    self.tm.depolarizeCells(activeCellsExternalApical=apicalInput,
-                            learn=learn)
-    self.tm.activateCells(activeColumns,
-                          reinforceCandidatesExternalApical=apicalInput,
-                          growthCandidatesExternalApical=apicalInput,
-                          learn=learn)
+    self.tm.depolarizeCells(learn=learn)
+    self.tm.activateCells(sorted(activeColumns), learn=learn)
 
 
   def reset(self):
